Better deploy

+++ post-install.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 git pull
-composer  install --no-dev --no-interaction --prefer-dist --optimize-autoloader
-
+if [ -f .vendor/bin/composer];then
+  ./vendor/bin/composer install --no-dev --no-interaction --prefer-dist --optimize-autoloader
+else
+  composer install --no-dev --no-interaction --prefer-dist --optimize-autoloader
+fi
 if [ -f artisan ]; then
     php artisan migrate --force
     php artisan queue:restart
